+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Releases
 
 ## Next
+* FEAT: Now supports transforms with regular expressions (Starofall)
 
 ## v4.0.1
 * FIX: Corrected an issue introduced in 4.0.0 where Grunt transforms were not running on `out`, `outDir`, `reference`, `mapRoot`, or `sourceRoot`.  (#220 - thanks to paulgambrell and JoeFirebaugh for the report.)